fix(collection-seeder): use rsvA/rsvB organism slugs, fix test imports for renamed sources
