+++ collection-seeding/tests/test_resistance_mutations.py
@@ -1,4 +1,4 @@
-from sources.resistance_mutations import CovidResistanceMutationsSource, _mature_name
+from sources.covid_resistance_mutations import CovidResistanceMutationsSource, _mature_name
 
 
 def test_name():
